Fix OOM errors in PostgreSQL dialect
Partial fix for #34, since that issue seems to cover more than just
PostgreSQL.

Per multiple contributors to that issue, the changes required are:

- ensure autoCommit is set to false on the Connection
- ensure fetchSize is set to a non-zero value on the PreparedStatement

In this patch I re-use the "batch.max.rows" property from
JdbcSourceConnectorConfig as the value for fetchSize. The rationale is
that the documented effect of that field is almost identical to the
documented effect of calling setFetchSize, i.e. it limits the number of
rows buffered internally. See
https://jdbc.postgresql.org/documentation/head/query.html#fetchsize-example.