CC-3638: Added logic to optionally quote tables and/or columns
Added the ability for JDBC source and sink connector configurations to explicitly set when table and column names would be quoted within DDL and DML statements, including the generated WHERE criteria. This is useful, for example, when column or table names don’t match exactly the records, and when incrementing and timestamp columns specified in the configuration include fully-qualified names, functions, or already-quoted values.

The `quote.table.names` and `quote.column.names` can be independently set to `always` or `never`. To maintain backward compatibility when existing configurations are unchanged, the defaults for the quoting behavior remains `always`. Therefore, to make use of this new feature requires explicitly setting them in connector configurations.