CC-7265: Support SQL Server DateTimeOffset (#750)
Added support for SQL Server’s DateTimeOffset type.

This change supports the jTDS driver that we ship by default, and even the latest version of this driver (which is pretty old) doesn't have support for DateTimeOffset. Per MS documentation, the dialect handles DateTimeOffset values by getting the string form of the DateTimeOffset and parsing them into a Timestamp.

The MS SQL Server JDBC driver does support the DateTimeOffset type, so when this driver is used the dialect simply delegates to the driver to get the timestamp for the column.