Make sink metadata column ordering more consistent.
Currently the use of a HashMap to store the columns loses the ordering of the columns from the schema.
When the create table is generated, the order of the columns is unpredictable.
This change makes the hash a LinkedHashMap which mostly preserves the original column order (with the exception that key columns are placed before non-key columns).