CC-8087: Remove plaintext logging of task configs
[JIRA](https://confluentinc.atlassian.net/browse/CC-8087)

These changes remove the logging of task configurations in plaintext, which should be done to avoid leaking database credentials and other secrets. Additionally, the log messages for whether a custom query has been specified by the user are flipped to correctly correspond to the actual presence of a custom query in the connector configuration.