CC-8898: Add better error messages when consumed records don’t match connector config
Added record validation that checks records key, key schema, value, and value schema to ensure that they match with the `delete.enabled` and `pk.mode` connector configuration properties.

Added unit tests that check all permutations of records with combinations of these configuration properties. Importantly, all of these unit tests were developed initially to identify the current behavior without the new validation, and thus asserted each permutation was either successfully written to a database or failed with one of the following:

* ConnectException (e.g., “PK mode for table 'dummy' is RECORD_KEY, but record key schema is missing”);
* NPEs; or
* SQLExceptions (e.g., “Values not bound to statement”)

Then when validation was added, the tests' assertions remained the same, except that the exceptions thrown in the failure cases are always ConnectException with much more useful error messages; NPEs and the aforementioned SQLException no longer occur.

The more useful error messages are specific to missing keys, key schemas, values, or value schemas, but are generally of the form:

> The JDBC sink connector 'my-connector' is configured with 'delete.enabled=true' and 'pk.mode=record_key' and therefore requires non-null record keys, but found record at (topic='dummy',partition=0,offset=1001,timestamp=null) with a null key and STRUCT key schema.