CC-9567: Allow users to define extra connection properties individually rather than only through JDBC URI
Extra connection properties are those that are prefixed with `connection.` and that are not already predefined by the connector’s ConfigDef. These properties are added by default in the GenericDatabaseDialect, since the existing `addConnectionProperties(…)` method was not explicitly overridden by any known connector — and if any did but did not call `super. addConnectionProperties(…)`, they just wouldn’t get this feature by default.

Modified one of the MySQL integration tests to pass a connection property via the extra properties rather than via the JDBC URI.