Update GenericDatabaseDialect.java
This plug-in does not limit the length of decimal data in the schema encapsulation. It is recommended to perfect the schema information according to the metadata information of the table during the schema encapsulation to ensure that the length of Decimal is less than 38 bits, which will also cause hive to be unable to parse the data type of Decimal.